@@ -19,18 +19,19 @@
 		"mdsvex": "^0.11.0",
 		"sass": "^1.75.0",
 		"supabase": ">=1.8.1",
-		"svelte": "^4.2.15",
+		"svelte": "^5.0.0-next.264",
 		"svelte-check": "^3.6.0",
 		"svelte-preprocess": "^5.1.4",
 		"tslib": "^2.4.1",
 		"typescript": "^5.0.0",
-		"vite": "^5.0.3"
+		"vite": "5.4.6"
 	},
 	"type": "module",
 	"dependencies": {
 		"@supabase/supabase-js": "^2.42.7",
 		"@types/jsdom": "^21.1.6",
 		"@vercel/analytics": "^1.3.1",
+		"runed": "^0.15.3",
 		"tsx": "^4.11.0",
 		"xmldom": "^0.6.0"
 	}
